# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/card/create": {
      "post": {
        "tags": ["Cards"],
        "summary": "Criar carta",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["front", "back"],
              "properties": {
                "front": { "type": "string" },
                "back": { "type": "string" },
                "deck_id": { "type": "string" },
                "user_id": { "type": "string" },
                "audio": { "type": "string" },
                "card_type": { "type": "string", "enum": ["text", "multiple_choice", "image"] },
                "options": { "type": "array", "items": { "type": "string" } },
                "correct_index": { "type": "integer", "minimum": 0, "maximum": 3 },
                "image": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "201": { "description": "Carta criada" }, "400": { "description": "Campos obrigatórios conforme o tipo da carta" } }
      }
    },
    "/card/create_card_in_lots": {
      "post": {
        "tags": ["Cards"],
        "summary": "Criar cartas em lote",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["cards", "deck_name"],
              "properties": {
                "cards": { "type": "array", "items": { "type": "object" } },
                "deck_name": { "type": "string" },
                "image": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "201": { "description": "Cartas criadas" }, "400": { "description": "cards obrigatório" } }
      }
    },
...
    "/card/get_cards_by_deck/{deck_id}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Cartas de um baralho",
        "parameters": [{ "name": "deck_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Lista de cartas" }, "400": { "description": "Erro" } }
      }
    },
    "/card/{card_id}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Buscar carta por ID",
        "parameters": [{ "name": "card_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Carta" }, "404": { "description": "Não encontrada" } }
      },
      "put": {
        "tags": ["Cards"],
        "summary": "Atualizar carta",
        "parameters": [
          { "name": "card_id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "schema": { "type": "object" } }
        ],
        "responses": { "200": { "description": "Carta atualizada" }, "404": { "description": "Não encontrada" } }
      },
      "delete": {
        "tags": ["Cards"],
        "summary": "Excluir carta",
        "parameters": [{ "name": "card_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Excluída" }, "404": { "description": "Não encontrada" } }
      }
    },
...
    "/billing/checkout": {
      "post": {
        "tags": ["Billing"],
        "summary": "Checkout nativo Asaas (PIX ou cartão) em web, Android e iOS",
        "security": [{ "Bearer": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "required": ["product_type", "product_id", "billing_type", "cpf_cnpj"],
            "properties": {
              "product_type": { "type": "string", "enum": ["plan", "book", "bundle", "course", "classroom"] },
              "product_id": { "type": "string" },
              "platform": { "type": "string", "enum": ["web", "android", "ios"] },
              "billing_type": { "type": "string", "enum": ["PIX", "CREDIT_CARD"] },
              "coupon_code": { "type": "string" },
              "cpf_cnpj": { "type": "string" },
              "credit_card": {
                "type": "object",
                "properties": {
                  "holder_name": { "type": "string" },
                  "number": { "type": "string" },
                  "expiry_month": { "type": "string" },
                  "expiry_year": { "type": "string" },
                  "ccv": { "type": "string" },
                  "postal_code": { "type": "string" },
                  "address_number": { "type": "string" },
                  "phone": { "type": "string" }
                }
              }
            }
          }
        }],
        "responses": { "200": { "description": "PIX (QR + payload) ou cobrança de cartão" } }
      }
    },
    "/billing/public/checkout": {
      "post": {
        "tags": ["Billing"],
        "summary": "Checkout público da turma (sempre visitante). Usuário novo: senha inicial = CPF, deve alterar no primeiro acesso.",
        "parameters": [{
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "required": ["product_type", "product_id", "email", "billing_type", "cpf_cnpj"],
            "properties": {
              "product_type": { "type": "string", "enum": ["classroom", "course"] },
              "product_id": { "type": "string" },
              "name": { "type": "string" },
              "email": { "type": "string" },
              "billing_type": { "type": "string", "enum": ["PIX", "CREDIT_CARD"] },
              "cpf_cnpj": { "type": "string" },
              "credit_card": { "type": "object" }
            }
          }
        }],
        "responses": { "200": { "description": "PIX ou URL de cartão. Perfil criado ou vinculado por e-mail/CPF. Sem senha." } }
      }
    },
    "/billing/public/payments/{payment_id}/sync": {
      "post": {
        "tags": ["Billing"],
        "summary": "Sincronizar pagamento do checkout público (e-mail + CPF, sem login)",
        "parameters": [
          { "name": "payment_id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["email", "cpf_cnpj"],
              "properties": {
                "email": { "type": "string" },
                "cpf_cnpj": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "200": { "description": "Status e granted" } }
      }
    },
    "/billing/public/payments/{payment_id}/pix": {
      "post": {
        "tags": ["Billing"],
        "summary": "QR PIX do checkout público (e-mail + CPF, sem login)",
        "parameters": [
          { "name": "payment_id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["email", "cpf_cnpj"],
              "properties": {
                "email": { "type": "string" },
                "cpf_cnpj": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "200": { "description": "encoded_image, payload e expiration_date" } }
      }
    },
    "/billing/me": {
      "get": {
        "tags": ["Billing"],
        "summary": "Assinatura e entitlements do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Estado da assinatura" } }
      }
    },
    "/billing/payments": {
      "get": {
        "tags": ["Billing"],
        "summary": "Histórico de pagamentos do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
    "/billing/payments/{payment_id}/sync": {
      "post": {
        "tags": ["Billing"],
        "summary": "Sincronizar status de um pagamento Asaas (PIX/cartão)",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "payment_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Status e granted" } }
      }
    },
    "/billing/payments/{payment_id}/pix": {
      "get": {
        "tags": ["Billing"],
        "summary": "Obter ou renovar QR Code PIX",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "payment_id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "encoded_image, payload e expiration_date" } }
      }
    },
    "/billing/update-payment": {
      "post": {
        "tags": ["Billing"],
        "summary": "Atualizar cartão da assinatura Asaas (ou manage_url legado do Play)",
        "security": [{ "Bearer": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "properties": {
              "cpf_cnpj": { "type": "string" },
              "credit_card": { "type": "object" }
            }
          }
        }],
        "responses": { "200": { "description": "Cartão atualizado" } }
      }
    },
    "/billing/cancel": {
      "post": {
        "tags": ["Billing"],
        "summary": "Cancelar assinatura atual",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Cancelada" } }
      }
    },
    "/billing/change-plan": {
      "post": {
        "tags": ["Billing"],
        "summary": "Upgrade ou downgrade de plano (mesmo provedor)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Plano atualizado" } }
      }
    },
...
    "/billing/google/verify": {
      "post": {
        "tags": ["Billing"],
        "summary": "Validar compra Google Play (assinaturas legadas)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Compra validada" } }
      }
    },
...
    "/admin/billing/payments": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar compras e transações",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
...
  }
}